chore: restructure core files
- Put tests in to their own directory
- Update src component names
- Update jest config to support aliases
- Update typescript files to support aliases
