+++ package.json
@@ -32,13 +32,17 @@
   ],
   "scripts": {
     "prebuild": "yarn typecheck && yarn clean",
-    "build": "tsc --outDir . --module esnext && tsc --outDir cjs --module commonjs",
-    "clean": "rimraf cjs react/*.ts react/*.js react/*.map index.d.ts index.test.d.ts",
+    "build": "tsc",
+    "clean": "rimraf dist coverage",
     "eslint": "eslint --fix '*.json' 'src/**/*.{ts,tsx}'",
     "format": "prettier '**/*.{md,js,jsx,json,ts,tsx}' --write",
     "prepare": "husky install",
     "prepublish": "yarn build",
-    "test": "jest --env=jsdom",
+    "pretest": "yarn typecheck",
+    "test": "jest",
+    "test:ci": "jest",
+    "test:coverage:watch": "jest --watch",
+    "test:dev": "jest --watch --no-coverage",
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
